+++ tests/Support/Objects/ExamplesPaginationQuery.php
@@ -35,6 +35,7 @@ class ExamplesPaginationQuery extends Query
 
     public function resolve($root, $args, $context, ResolveInfo $resolveInfo, Closure $getSelectFields): LengthAwarePaginator
     {
+        /** @var array<int,array{test:string}> $data */
         $data = include __DIR__ . '/data.php';
 
         $take = $args['take'];