+++ CHANGELOG.md
@@ -4,6 +4,9 @@ CHANGELOG
 [Next release](https://github.com/rebing/graphql-laravel/compare/8.2.1...master)
 --------------
 
+### Removed
+- Support for PHP 7.2, PHP 7.3 and Laravel 7.0 (all EOL) [\#914 / mfn](https://github.com/rebing/graphql-laravel/pull/914)
+
 2022-01-30, 8.2.1
 -----------------
 ### Fixed